Interfaces are moved to original packages
