+++ kotlin-emotion/src/main/kotlin/emotion/css/CSSStyleSheet.kt
@@ -1,5 +1,7 @@
 package emotion.css
 
+import emotion.sheet.StyleSheet
+
 external interface CSSStyleSheet : StyleSheet {
     fun speedy(value: Boolean)
 }
