+++ kotlin-emotion/src/main/kotlin/emotion/css/cache.kt
@@ -3,4 +3,6 @@
 
 package emotion.css
 
+import emotion.cache.EmotionCache
+
 external val cache: EmotionCache
