+++ kotlin-emotion/src/main/kotlin/emotion/sheet/StyleSheet.kt
@@ -1,4 +1,4 @@
-package emotion.css
+package emotion.sheet
 
 import kotlinext.js.ReadonlyArray
 import org.w3c.dom.HTMLElement
